requirements.txt
# Machine Learning
pandas
numpy
joblib
xgboost
seaborn
scikit-learn
imbalanced-learn

# API
flask
python-dotenv

# Dashboard
streamlit
requests

# Interprétabilité (optionnel)
shap
matplotlib










